complete drawmanager refactoring phase for now
